ci: introduce changelog generation with git-chglog (#13)
* ci: introduce changelog generation with git-chglog
It has been bugging me for some time now that releases of the Helm Chart are quite ambiguous.
Future Helm Chart releases will include a link to the latest changelog.
Users will now also be informed about unreleased features which will make it into the upcoming release.
Additionally, if I don't want to blow up release changelogs in the future, I will need to squash commits more often.

This is an outcome of #9. I'm not 100% happy with it at the moment but it eases my mind about releases.

Signed-off-by: Torben Dury <torbendury@googlemail.com>

* feat: shorten time interval for liveness and readiness probe (#12)

The liveness probe used to only check every 5 seconds if the container is able to serve traffic.
This lead to the fact that Kubernetes would only stop traffic to the container after ~10 seconds which is unacceptable.
Also, the readiness probe now fails faster and checks in a shorter interval.

This closes #10.

Signed-off-by: Torben Dury <torbendury@googlemail.com>

* Bump appVersion to e8e5b7b

* chore(core): move admission UID retrieval (#14)

also correct minor documentation issue.
The UID retrieval is most probably optimized away on build time but it makes it more clear
where the UID is actually being used. No need for a separate variable.

Signed-off-by: Torben Dury <torbendury@googlemail.com>

* ci: introduce changelog generation with git-chglog
It has been bugging me for some time now that releases of the Helm Chart are quite ambiguous.
Future Helm Chart releases will include a link to the latest changelog.
Users will now also be informed about unreleased features which will make it into the upcoming release.
Additionally, if I don't want to blow up release changelogs in the future, I will need to squash commits more often.

This is an outcome of #9. I'm not 100% happy with it at the moment but it eases my mind about releases.

Signed-off-by: Torben Dury <torbendury@googlemail.com>

* chore: CHANGELOG

Signed-off-by: Torben Dury <torbendury@googlemail.com>

* doc: refer to changelog

Signed-off-by: Torben Dury <torbendury@googlemail.com>

---------

Signed-off-by: Torben Dury <torbendury@googlemail.com>
Co-authored-by: GitHub Actions <github-actions@github.com>